testapp/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
boto3==1.35.66
botocore==1.35.66
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Deprecated==1.2.14
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.5
fastapi-cli==0.0.5
googleapis-common-protos==1.66.0
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
opentelemetry-api==1.28.1
opentelemetry-distro==0.49b1
opentelemetry-exporter-otlp==1.28.1
opentelemetry-exporter-otlp-proto-common==1.28.1
opentelemetry-exporter-otlp-proto-grpc==1.28.1
opentelemetry-exporter-otlp-proto-http==1.28.1
opentelemetry-instrumentation==0.49b1
opentelemetry-instrumentation-asgi==0.49b1
opentelemetry-instrumentation-fastapi==0.49b1
opentelemetry-proto==1.28.1
opentelemetry-sdk==1.28.1
opentelemetry-semantic-conventions==0.49b1
opentelemetry-util-http==0.49b1
packaging==24.2
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.2
typer==0.13.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
wrapt==1.16.0
zipp==3.21.0
